+++ src/cpp/viper/utils/FIR.h
@@ -7,7 +7,7 @@ public:
     FIR();
     ~FIR();
 
-    void FilterSamples(int *samples, uint32_t size);
+    void FilterSamples(float *samples, uint32_t size);
     void FilterSamplesInterleaved(float *samples, uint32_t size, uint32_t channels);
     int GetBlockLength();
     int LoadCoefficients(const float *coeffs, uint32_t coeffsize, int blockLength);
